+++ packages/expo-navigation-bar/src/ExpoNavigationBar.ts
@@ -1,10 +1,6 @@
 import type { EventSubscription } from 'expo-modules-core';
 
-import type {
-  NavigationBarButtonStyle,
-  NavigationBarVisibility,
-  NavigationBarVisibilityEvent,
-} from './NavigationBar.types';
+import type { NavigationBarVisibility, NavigationBarVisibilityEvent } from './NavigationBar.types';
 
 export default {} as {
   addListener: (
@@ -12,7 +8,7 @@ export default {} as {
     listener: (event: NavigationBarVisibilityEvent) => void
   ) => EventSubscription;
 
-  setButtonStyleAsync: (style: NavigationBarButtonStyle) => Promise<void>;
+  setButtonStyleAsync: (style: 'light' | 'dark') => Promise<void>;
 
   setVisibilityAsync: (visibility: NavigationBarVisibility) => Promise<void>;
   getVisibilityAsync: () => Promise<NavigationBarVisibility>;