+++ packages/expo-router/src/views/Navigator.tsx
@@ -2,7 +2,6 @@
 'use client';
 
 import * as React from 'react';
-import { isEdgeToEdge } from 'react-native-is-edge-to-edge';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import { Screen } from './Screen';
@@ -161,7 +160,7 @@ function NavigatorSlot() {
  * The default navigator for the app when no root _layout is provided.
  */
 export function DefaultNavigator() {
-  if (process.env.EXPO_OS === 'android' && isEdgeToEdge()) {
+  if (process.env.EXPO_OS === 'android') {
     return <SlotNavigator />;
   }
   return (