+++ packages/expo-router/src/views/Sitemap.tsx
@@ -3,16 +3,7 @@
 
 import Constants from 'expo-constants';
 import React from 'react';
-import {
-  Image,
-  StyleSheet,
-  Text,
-  View,
-  ScrollView,
-  Platform,
-  StatusBar,
-  ViewStyle,
-} from 'react-native';
+import { Image, StyleSheet, Text, View, ScrollView, Platform, ViewStyle } from 'react-native';
 import { SafeAreaView } from 'react-native-safe-area-context';
 
 import { NoSSR } from './NoSSR';
@@ -20,7 +11,6 @@ import { Pressable, PressableProps } from './Pressable';
 import { useSitemap, SitemapType } from './useSitemap';
 import { Link } from '../link/Link';
 import { NativeStackNavigationOptions } from '../react-navigation/native-stack';
-import { canOverrideStatusBarBehavior } from '../utils/statusbar';
 
 const INDENT = 20;
 
@@ -79,7 +69,6 @@ function SitemapInner() {
   const Wrapper = Platform.OS === 'android' ? SafeAreaView : View;
   return (
     <Wrapper style={styles.container} testID="expo-router-sitemap">
-      {canOverrideStatusBarBehavior && <StatusBar barStyle="light-content" />}
       <ScrollView
         contentContainerStyle={styles.scroll}
         automaticallyAdjustContentInsets